feat: Module System - Implement module core - Add custom annotations - Refactor demo plugin to use module system - Code cleanup
