+++ paper/src/main/java/xyz/femdev/femutils/paper/gui/ClickContext.java
@@ -7,10 +7,10 @@ import org.bukkit.inventory.Inventory;
 /**
  * Provides context for a GUI click event.
  *
- * @param player     the clicking player
- * @param inventory  the inventory being interacted with
- * @param slot       the clicked slot index
- * @param clickType  the type of click performed
+ * @param player    the clicking player
+ * @param inventory the inventory being interacted with
+ * @param slot      the clicked slot index
+ * @param clickType the type of click performed
  */
 public record ClickContext(Player player, Inventory inventory, int slot, ClickType clickType) {
 }
